SIDM-1275: Open up Privacy Policy and T&C pages in the same browser window. (#52)
https://tools.hmcts.net/jira/browse/SIDM-1275
The behaviour of the links to Privacy Policy page and to T&C on the Sign Up page is not consistent - the links in the footer open up in the same browser window while the links close to the CTA button open up in a blank window. This PR fixes the inconsistency and as a side effect enables the "Back" button on the Privacy Policy page to work properly.